requirements.txt
streamlit==1.32.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
pandas==2.2.1
matplotlib==3.8.3
plotly==5.19.0
openpyxl==3.1.2
